+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- ### Fixed -->
 <!-- ### Security -->
 
+## [Unreleased]
+### Changed
+- Global config file on MacOS now follows XDG standard so its location can be changed by setting `XDG_CONFIG_HOME`
+
+
 ## v0.7.1 (2019-01-14)
 ### Added
 - improved help when using `-h` and `--help` tags